feat: checkbox & multiple

+++ my-react-app/src/components/dynamicform/altertable.js
@@ -17,6 +17,7 @@ function AlterTable({ onAddColumn }) {
   const [modifiedColumnTitle, setModifiedColumnTitle] = useState("");
   const [isTitleModalOpen, setIsTitleModalOpen] = useState(false);
   const [selectedColumn, setSelectedColumn] = useState(null);
+  const [newColumnInputType, setNewColumnInputType] = useState("default");
 
   const handleModalOpen = () => {
     setIsModalOpen(true);
@@ -74,14 +75,21 @@ function AlterTable({ onAddColumn }) {
   };
 
   const handleDeleteTable = async (tableName) => {
-    try {
-      await axios.delete(`${apiURL}/get_table_structure/${tableName}`);
-      const updatedMatchingTables = matchingTables.filter(
-        (table) => table !== tableName
-      );
-      setMatchingTables(updatedMatchingTables);
-    } catch (error) {
-      console.error("Error deleting table:", error);
+    // Display a confirmation dialog
+    const confirmDelete = window.confirm(
+      "Are you sure you want to delete this table?"
+    );
+
+    if (confirmDelete) {
+      try {
+        await axios.delete(`${apiURL}/get_table_structure/${tableName}`);
+        const updatedMatchingTables = matchingTables.filter(
+          (table) => table !== tableName
+        );
+        setMatchingTables(updatedMatchingTables);
+      } catch (error) {
+        console.error("Error deleting table:", error);
+      }
     }
   };
 
@@ -105,36 +113,61 @@ function AlterTable({ onAddColumn }) {
   };
 
   const handleDropColumn = async (columnName) => {
-    try {
-      await axios.patch(`${apiURL}/get_table_structure/${tableName}/`, {
-        column_name: columnName,
-      });
-      const updatedColumns = tableColumns.filter(
-        (column) => column.name !== columnName
-      );
-      setTableColumns(updatedColumns);
-    } catch (error) {
-      console.error("Error dropping column:", error);
+    // Display a confirmation dialog
+    const confirmDrop = window.confirm(
+      "Are you sure you want to drop this column?"
+    );
+
+    if (confirmDrop) {
+      try {
+        await axios.patch(`${apiURL}/get_table_structure/${tableName}/`, {
+          column_name: columnName,
+        });
+        const updatedColumns = tableColumns.filter(
+          (column) => column.name !== columnName
+        );
+        setTableColumns(updatedColumns);
+      } catch (error) {
+        console.error("Error dropping column:", error);
+      }
     }
   };
 
   const handleAddColumn = async () => {
     try {
-      const randomNumber = Math.floor(Math.random() * 1000000);
-      const newColumnName = `new_column_${randomNumber}`;
+      const min = 100;
+      const max = 1000000;
+      const randomNumber = Math.floor(Math.random() * (max - min) + min);
+      const newColumnName = `col_${randomNumber}`;
+
+      let columnOptions = undefined; // Initialize columnOptions variable
+      let columnDataType = newColumnType; // Initialize columnDataType variable
+
+      if (
+        newColumnType === "DROPDOWN" ||
+        newColumnType === "MULTIPLESELECT" ||
+        newColumnType === "CHECKBOX"
+      ) {
+        // If column type requires options, convert comma-separated options to an array
+        columnOptions = newColumnOptions
+          .split(",")
+          .map((option) => option.trim());
+        columnDataType = "ENUM"; // Set column data type to ENUM
+      }
+
+      console.log("Received column data type:", newColumnType);
+      console.log("Received input type:", newColumnInputType); // Log the newColumnInputType
+      console.log("Generated column data type:", columnDataType);
+
       const response = await axios.put(
         `${apiURL}/get_table_structure/${tableName}/`,
         {
           column_name: newColumnName,
-          data_type: newColumnType,
+          data_type: columnDataType,
+          enum_type: newColumnInputType, // Use newColumnInputType here
           nullable: !isRequired, // Invert isRequired to set nullable status
           name: newColumnTitle,
-          options:
-            newColumnType === "DROPDOWN" ||
-            newColumnType === "MULTIPLESELECT" ||
-            newColumnType === "CHECKBOX"
-              ? newColumnOptions
-              : undefined,
+          options: columnOptions, // Pass the array directly
           width: newColumnWidth,
           required: isRequired,
         }
@@ -168,9 +201,26 @@ function AlterTable({ onAddColumn }) {
     return title;
   };
 
-  const handleCheckboxChange = (event) => {
-    setIsRequired(event.target.checked);
-    setNewColumnTitle(updateColumnTitle(newColumnTitle, event.target.checked));
+  const handleCheckboxChange = (event, columnName) => {
+    const checkedValue = event.target.checked;
+    const optionValue = event.target.value;
+    const currentFormData = { ...formData };
+
+    if (!checkedValue) {
+      // Remove the unchecked option from the form data
+      currentFormData[columnName] = currentFormData[columnName].filter(
+        (option) => option !== optionValue
+      );
+    } else {
+      // Add the checked option to the form data
+      currentFormData[columnName] = [
+        ...(currentFormData[columnName] || []),
+        optionValue,
+      ];
+    }
+
+    // Update the form data state
+    setFormData(currentFormData);
   };
 
   const handleTitleChange = (event) => {
@@ -218,6 +268,47 @@ function AlterTable({ onAddColumn }) {
       }
     } catch (error) {
       console.error("Error updating column title:", error);
+    }
+  };
+
+  const handleColumnTypeChange = (event) => {
+    const selectedColumnType = event.target.value;
+    setNewColumnType(selectedColumnType);
+
+    // Automatically set input type based on column type
+    switch (selectedColumnType) {
+      case "VARCHAR(250)":
+      case "TEXT":
+      case "TEXTAREA":
+      case "FILE":
+      case "IMAGE":
+        setNewColumnInputType("default");
+        break;
+      case "DECIMAL":
+      case "INTEGER":
+        setNewColumnInputType("number");
+        break;
+      case "BOOLEAN":
+        setNewColumnInputType("checkbox");
+        break;
+      case "TIMESTAMP":
+        setNewColumnInputType("datetime-local");
+        break;
+      case "DROPDOWN":
+        setNewColumnInputType("dropdown");
+        break;
+      case "MULTIPLESELECT":
+        setNewColumnInputType("multiple_select");
+        break;
+      case "CHECKBOX":
+        setNewColumnInputType("checkbox");
+        break;
+      case "ENUM":
+        // If the column data type is ENUM, set the input type based on the received input type
+        setNewColumnInputType(newColumnInputType); // Use the received input type
+        break;
+      default:
+        setNewColumnInputType("default");
     }
   };
 
@@ -377,23 +468,55 @@ function AlterTable({ onAddColumn }) {
                       column.width || ""
                     } focus:outline-none focus:border-blue-500`}
                   />
-                ) : column.type === "DROPDOWN" ||
-                  column.type === "MULTIPLESELECT" ||
-                  column.type === "CHECKBOX" ? (
-                  <select
-                    value={formData[column.name] || ""}
-                    onChange={(event) => handleInputChange(event, column.name)}
-                    className={`border border-gray-300 rounded px-4 py-2 w-full ${
-                      column.width || ""
-                    } focus:outline-none focus:border-blue-500`}
-                  >
-                    <option value="">Select</option>
-                    {column.options.map((option, index) => (
-                      <option key={index} value={option}>
-                        {option}
-                      </option>
+                ) : column.type === "DROPDOWN" ? (
+                  <div className="mb-4">
+                    <label className="block mb-1 text-gray-700">
+                      Options (comma separated)
+                    </label>
+                    <input
+                      type="text"
+                      value={newColumnOptions}
+                      onChange={(e) => setNewColumnOptions(e.target.value)}
+                      className="border border-gray-300 rounded px-4 py-2 w-full focus:outline-none focus:border-blue-500"
+                    />
+                  </div>
+                ) : column.type === "MULTIPLESELECT" ? (
+                  <div className="mb-4">
+                    <label className="block mb-1 text-gray-700">
+                      Options (comma separated)
+                    </label>
+                    <input
+                      type="text"
+                      value={newColumnOptions}
+                      onChange={(e) => setNewColumnOptions(e.target.value)}
+                      className="border border-gray-300 rounded px-4 py-2 w-full focus:outline-none focus:border-blue-500"
+                    />
+                  </div>
+                ) : column.enum_type === "CHECKBOX" ? (
+                  <div className="mb-4">
+                    <label className="block mb-1 text-gray-700">
+                      {column.column_fullname}
+                    </label>
+                    {column.options.map((option, optionIndex) => (
+                      <div key={optionIndex}>
+                        <input
+                          type="checkbox"
+                          id={`${column.name}_${optionIndex}`}
+                          value={option}
+                          checked={
+                            formData[column.name] &&
+                            formData[column.name].includes(option)
+                          }
+                          onChange={(e) => handleCheckboxChange(e, column.name)}
+                          className="mr-2"
+                        />
+
+                        <label htmlFor={`${column.name}_${optionIndex}`}>
+                          {option}
+                        </label>
+                      </div>
                     ))}
-                  </select>
+                  </div>
                 ) : (
                   <input
                     type="text"
@@ -404,6 +527,7 @@ function AlterTable({ onAddColumn }) {
                     } focus:outline-none focus:border-blue-500`}
                   />
                 )}
+
                 <button
                   type="button"
                   className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded mr-2"
@@ -428,10 +552,11 @@ function AlterTable({ onAddColumn }) {
               </label>
               <select
                 value={newColumnType}
-                onChange={(e) => setNewColumnType(e.target.value)}
+                onChange={handleColumnTypeChange} // Call handleColumnTypeChange when new column type changes
                 className="border border-gray-300 rounded px-4 py-2 w-full focus:outline-none focus:border-blue-500"
               >
                 <option value="VARCHAR(250)">Text</option>
+                <option value="TEXT">Text</option>
                 <option value="TEXTAREA">Text Area</option>
                 <option value="DECIMAL">Decimal</option>
                 <option value="INTEGER">Integer</option>
@@ -444,6 +569,7 @@ function AlterTable({ onAddColumn }) {
                 <option value="CHECKBOX">Checkbox</option>
               </select>
             </div>
+
             {/* <div className="mb-4">
               <label className="block mb-1 text-gray-700">
                 New Column Title
@@ -526,6 +652,27 @@ function AlterTable({ onAddColumn }) {
                     className="border border-gray-300 rounded px-4 py-2 w-full focus:outline-none focus:border-blue-500"
                   />
                 </div>
+                {
+                  // Check if the newColumnType is one of "DROPDOWN", "MULTIPLESELECT", or "CHECKBOX"
+                  ["DROPDOWN", "MULTIPLESELECT", "CHECKBOX"].includes(
+                    newColumnType
+                  ) && (
+                    // If it is, render the following JSX
+                    <div className="mb-4">
+                      <label className="block mb-1 text-gray-700">
+                        Options (comma separated)
+                      </label>
+                      {/* Input field for entering options */}
+                      <input
+                        type="text"
+                        value={newColumnOptions}
+                        onChange={(e) => setNewColumnOptions(e.target.value)}
+                        className="border border-gray-300 rounded px-4 py-2 w-full focus:outline-none focus:border-blue-500"
+                      />
+                    </div>
+                  )
+                }
+
                 <div className="mb-4">
                   <label className="block mb-1 text-gray-700">
                     Column Width
